[HW][Comb][ExportVerilog] Fix HWIntegerType's C++ type to allow hw::IntType.
Previously, the dialect type HWIntegerType had a declared C++ type of
mlir::IntegerType. This was fine in the past, when HWIntegerType
was just a lightweight wrapper around the MLIR type. However, this
gradually became incorrect when we created !hw.int for holding
hardware integers with parametric width and as we started using !hw.int
in existing ops that were declared to have HWIntegerType operands and
results.

Although it was technically incorrect, this didn't cause issues because:

1. Parametric !hw.int isn't really fully supported, and most existing
   passes assume we have statically-known widths.
2. Because the C++ type was incorrect, most of the code implicictly
   converted the operand/result types to mlir::IntegerType, which would
   work fine as long as the runtime type was indeed mlir::IntegerType.

2) is particularly pernicious, since it masked places where we should
have been explicitly casting values to mlir::IntegerType, since the
compiler would assume that it should implicitly convert to
mlir::IntegerType without considering the possibility that we actually
have an hw::IntType.

We need to actually fix this now that upstream MLIR introduced a
TypedValue type, since TypedValue will perform a runtime assertion to
check that the wrapped value is indeed of the correct type.
Specifically, ops that have a dialect type of HWIntegerType would
declare a C++ type of mlir::IntegerType, and if the op _actually_ had a
type of hw::Int, the new TypedValue would fail the runtime type check.

This commit changes HWIntegerType's C++ type to be
TypeVariant<mlir::IntegerType, hw::IntType>, which correct accounts for
both possible cases, and it allows for the TypedValue runtime type
checks to succeed.

This change has the side effect of forcing us to explicitly cast most
ops' results to mlir::IntegerType in places where we currently assume
that the type is a non-parametric hardware integer. Ideally, we should
be checking whether an op result is parametric before we run folds that
rely on having a statically-known width, but putting in the explicit
casts preserves the existing behavior while at least making it more
explicit that we need an integer of statically-known width.

Update tablegen.

+++ lib/Dialect/Comb/CombOps.cpp
@@ -182,7 +182,7 @@ LogicalResult ReplicateOp::verify() {
   // The source must be equal or smaller than the dest type, and an even
   // multiple of it.  Both are already known to be signless integers.
   auto srcWidth = getOperand().getType().cast<IntegerType>().getWidth();
-  auto dstWidth = getType().getWidth();
+  auto dstWidth = getType().cast<IntegerType>().getWidth();
   if (srcWidth == 0)
     return emitOpError("replicate does not take zero bit integer");
 
@@ -241,7 +241,7 @@ static unsigned getTotalWidth(ValueRange inputs) {
 }
 
 LogicalResult ConcatOp::verify() {
-  unsigned tyWidth = getType().getWidth();
+  unsigned tyWidth = getType().cast<IntegerType>().getWidth();
   unsigned operandsTotalWidth = getTotalWidth(getInputs());
   if (tyWidth != operandsTotalWidth)
     return emitOpError("ConcatOp requires operands total width to "
@@ -277,7 +277,7 @@ LogicalResult ConcatOp::inferReturnTypes(MLIRContext *context,
 
 LogicalResult ExtractOp::verify() {
   unsigned srcWidth = getInput().getType().cast<IntegerType>().getWidth();
-  unsigned dstWidth = getType().getWidth();
+  unsigned dstWidth = getType().cast<IntegerType>().getWidth();
   if (getLowBit() >= srcWidth || srcWidth - getLowBit() < dstWidth)
     return emitOpError("from bit too large for input"), failure();
